refactor(server): isolate ifthenelse pandas implem
