+++ server/tests/steps/test_ifthenelse.py
@@ -2,7 +2,8 @@ import pytest
 from pandas import DataFrame
 
 from tests.utils import assert_dataframes_equals
-from weaverbird.steps.ifthenelse import IfthenelseStep
+from weaverbird.backends.pandas_executor.steps.ifthenelse import execute_ifthenelse
+from weaverbird.pipeline.steps.ifthenelse import IfthenelseStep
 
 
 @pytest.fixture
@@ -11,7 +12,7 @@ def sample_df() -> DataFrame:
 
 
 def test_simple_condition(sample_df):
-    result_df = IfthenelseStep(
+    step = IfthenelseStep(
         **{
             'name': 'ifthenelse',
             'newColumn': 'result',
@@ -19,7 +20,8 @@ def test_simple_condition(sample_df):
             'then': 10,
             'else': 0,
         }
-    ).execute(sample_df)
+    )
+    result_df = execute_ifthenelse(step, sample_df)
 
     expected_df = DataFrame({'a_bool': [True, True, False], 'result': [10, 10, 0]})
 
@@ -28,7 +30,7 @@ def test_simple_condition(sample_df):
 
 def test_simple_condition_strings():
     sample_df = DataFrame({'a_str': ["test", "test", "autre chose"]})
-    result_df = IfthenelseStep(
+    step = IfthenelseStep(
         **{
             'name': 'ifthenelse',
             'newColumn': 'test',
@@ -36,7 +38,8 @@ def test_simple_condition_strings():
             'then': "\"foo\"",
             'else': "\"bar\"",
         }
-    ).execute(sample_df)
+    )
+    result_df = execute_ifthenelse(step, sample_df)
 
     expected_df = DataFrame(
         {'a_str': ["test", "test", "autre chose"], 'test': ["foo", "foo", "bar"]}
@@ -47,7 +50,7 @@ def test_simple_condition_strings():
 
 def test_then_should_support_formulas():
     base_df = DataFrame({'a_bool': [True, False, True], 'a_number': [1, 2, 3]})
-    result_df = IfthenelseStep(
+    step = IfthenelseStep(
         **{
             'name': 'ifthenelse',
             'newColumn': 'result',
@@ -55,7 +58,8 @@ def test_then_should_support_formulas():
             'then': 'a_number',
             'else': 'a_number * -1',
         }
-    ).execute(base_df)
+    )
+    result_df = execute_ifthenelse(step, base_df)
 
     expected_df = DataFrame(
         {'a_bool': [True, False, True], 'a_number': [1, 2, 3], 'result': [1, -2, 3]}
@@ -66,7 +70,7 @@ def test_then_should_support_formulas():
 
 def test_then_should_support_nested_else():
     base_df = DataFrame({'a_bool': [True, False, False], 'a_number': [1, 2, 3]})
-    result_df = IfthenelseStep(
+    step = IfthenelseStep(
         **{
             'name': 'ifthenelse',
             'newColumn': 'result',
@@ -82,7 +86,8 @@ def test_then_should_support_nested_else():
                 },
             },
         }
-    ).execute(base_df)
+    )
+    result_df = execute_ifthenelse(step, base_df)
 
     expected_df = DataFrame(
         {'a_bool': [True, False, False], 'a_number': [1, 2, 3], 'result': [3, 2, 1]}
@@ -103,7 +108,7 @@ def test_isnull():
         }
     )
 
-    result = step.execute(df)
+    result = execute_ifthenelse(step, df)
     assert_dataframes_equals(result, DataFrame({'a_bool': [True, False, None], 'test': [0, 0, 1]}))
 
 
@@ -118,4 +123,4 @@ def test_benchmark_ifthenelse(benchmark):
             "else": "0",
         }
     )
-    benchmark(step.execute, big_df)
+    benchmark(execute_ifthenelse, step, big_df)
