+++ server/weaverbird/backends/pandas_executor/steps/__init__.py
@@ -20,8 +20,8 @@ from .fillna import execute_fillna
 from .filter import execute_filter
 from .formula import execute_formula
 from .fromdate import execute_fromdate
+from .ifthenelse import execute_ifthenelse
 
-# from .ifthenelse import execute as execute_ifthenelse
 # from .join import execute as execute_join
 # from .lowercase import execute as execute_lowercase
 # from .moving_average import execute as execute_moving_average
@@ -65,7 +65,7 @@ steps_executors: Dict[str, StepExecutor] = {
     'filter': execute_filter,
     'formula': execute_formula,
     'fromdate': execute_fromdate,
-    # 'ifthenelse': execute_ifthenelse,
+    'ifthenelse': execute_ifthenelse,
     # 'join': execute_join,
     # 'lowercase': execute_lowercase,
     # 'moving_average': execute_moving_average,
